# src/main/java/br/com/intergado/projetobasico/dto/FazendaDTO.java
package br.com.intergado.projetobasico.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class FazendaDTO {
    private Long id;
    private String name;
    private List<AnimalDTO> animals;
}